--- server/go.mod
@@ -1,7 +1,0 @@
-module jsimmons/pokerserver
-go 1.20
-
-require (
-	github.com/gorilla/websocket v1.5.1 // indirect
-	golang.org/x/net v0.17.0 // indirect
-)
