--- server/readme.md
@@ -1,1 +1,0 @@
-run `air` to live reload and run
